Move label extraction from cron to pipeline task
Doing this task on as-needed basiss is much more efficient and scalable than batching
into a chron task. Also, this alleviates the issues with race conditions.
